Update Alternate Version Strategy
Updates alternate version strategy to drop a default package for each
Fedora release. Instead, the alternate will include symlinks that
provide the standard binaries in the usual place. This will allow only
one alternate to be installed. DNF will explicitly require a different
version to be removed first.

Two versions of the command line client will be produced. One is
standard sibling to other sub packages in the alternate version. The
other version of  the client includes environment module support and
will allow multiple clients on a host.